slight ammends to the main controller repo regarding id type

+++ backend/src/main/java/com/example/backend/repository/ServicesRepository.java
@@ -4,4 +4,4 @@ import org.springframework.stereotype.Repository;
 import com.example.backend.model.Services; 
 
 @Repository
-public interface ServicesRepository extends CrudRepository <Services, Long> { }
+public interface ServicesRepository extends CrudRepository <Services, Integer> { }
